requirements.txt
aiohttp==3.9.5
discord.py==2.4.0
psutil==6.0.0
python-dotenv==1.0.1
rcon==2.4.9
